requirements.txt
torch
scikit-learn
numpy
pandas
matplotlib
seaborn
scipy
biopython
ruamel.yaml
pyyaml
tqdm
optuna
tabulate
ipykernel
jupyter
tensorboard
transformers
sentencepiece
# esm?
# QUEDA PENDIENTE REVISAR ESTO!!